+++ apps/node-remote/module-federation.config.ts
@@ -1,0 +1,13 @@
+import { createModuleFederationConfig } from '@module-federation/rsbuild-plugin';
+
+export default createModuleFederationConfig({
+  name: 'node_remote',
+  filename: 'remoteEntry.js',
+  exposes: {
+    './test': './src/expose.js',
+  },
+  remotes: {
+    remote1: 'node_dynamic_remote@http://localhost:3026/remoteEntry.js',
+  },
+  dts: false,
+});