+++ apps/website-new/module-federation.config.ts
@@ -6,9 +6,10 @@ const LANGUAGES = ['zh', 'en'];
 
 const exposes = {
   // basic
-  [`./rspack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/rspack.mdx`,
-  [`./webpack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/webpack.mdx`,
-  [`./rspress-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/rspress.mdx`,
+  [`./plugins-overview-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/overview.mdx`,
+  [`./rspack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/rspack.mdx`,
+  [`./webpack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/webpack.mdx`,
+  [`./rspress-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/rspress.mdx`,
   [`./cli-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/cli.mdx`,
   [`./type-prompt-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/type-prompt.mdx`,
   [`./css-isolate-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/css-isolate.mdx`,
@@ -66,6 +67,7 @@ const exposes = {
 
   // performance
   [`./shared-tree-shaking-overview-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/performance/shared-treeshaking.mdx`,
+  [`./node-${LANGUAGE}`]: `./docs/${LANGUAGE}/blog/node.mdx`,
 };
 
 export default createModuleFederationConfig({