+++ packages/bridge/vue3-bridge/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/bridge-vue3
 
+## 0.23.0
+
+### Patch Changes
+
+- @module-federation/runtime@0.23.0
+- @module-federation/sdk@0.23.0
+- @module-federation/bridge-shared@0.23.0
+
 ## 0.22.1
 
 ### Patch Changes