+++ packages/rspress-plugin/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/rspress-plugin
 
+## 0.23.0
+
+### Patch Changes
+
+- Updated dependencies [3aa3cf9]
+  - @module-federation/rsbuild-plugin@0.23.0
+  - @module-federation/enhanced@0.23.0
+  - @module-federation/sdk@0.23.0
+  - @module-federation/error-codes@0.23.0
+
 ## 0.22.1
 
 ### Patch Changes